% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.agedepthmodel}
\alias{plot.agedepthmodel}
\title{Plot one or two age-depth models}
\usage{
\method{plot}{agedepthmodel}(
  x,
  y = NULL,
  colours = c("red", "green"),
  ageunits = "years bp",
  depthunits = "cm",
  include_input_ages = TRUE,
  include_labels = TRUE
)
}
\arguments{
\item{x}{An object of type \code{agedepthmodel}}

\item{y}{A second, optional, object of type \code{agedepthmodel}}

\item{colours}{The colours to be used for the two age-depth models}

\item{ageunits}{The description of the units to be used on the age axis}

\item{depthunits}{The description of the units to be used on the depth axis}

\item{include_input_ages}{Boolean. If true, plots the ages used as input to the model}

\item{include_labels}{Boolean. If true, include the observation labels from the input data}
}
\value{
An object of type \code{ggplot}
}
\description{
Plot of one or two age-depth models, including input and posterior ages where available.
}
\examples{
# plot(sample_data())

}
