% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict.agedepthmodel}
\alias{predict.agedepthmodel}
\title{Predict age for a given depth for an age-depth model}
\usage{
\method{predict}{agedepthmodel}(x, depth)
}
\arguments{
\item{x}{An object of type \code{agedepthmodel}}

\item{depth}{The depth at which the age is required}
}
\value{
A named vector containing the mean age according to the model, together with the minimum and maximum values (95\% HDI)
}
\description{
Use the input value to determine the age at a given depth.
}
