% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run an age-depth model}
\usage{
run_model(
  x,
  model = c("uniform_simple", "bacon_simple", "sequential_simple", "sequential_full",
    "uniform_full", "bacon_full"),
  ...
)
}
\arguments{
\item{x}{A data frame containing age depth data}

\item{model}{A string specifying the model to be run}

\item{chains}{The number of MCMC chains to be run}

\item{warmup}{The number of MCMC iterations to be discarded as warmup}

\item{iter}{The total number of MCMC iterations}

\item{modelled_points}{(Bacon only) the number of points to model. The number of calculated rates is one less that this number.}

\item{output_points}{(Bacon only) the number of depths to provide an age for in the output}

\item{a_alpha}{(Bacon only) the a input for the alpha (rate) prior}

\item{b_alpha}{(Bacon only) the b input for the alpha (rate) prior}

\item{a_w}{(Bacon only) the a input for the memory prior}

\item{b_w}{(Bacon only) the b input for the memory prior}

\item{t_df}{(Bacon only) the number of degrees of freedom for the t-distribution in the likelihood function}

\item{use_normal}{(Bacon only) If 1, use the normal distribution instead of the t-distribution in the likelihood}
}
\description{
Runs the specified age-depth model and returns an object of type \code{agedepthmodel}

The input data frame should contain four columns: \code{label}, a label for the observation, \code{age}, an uncalibrated C14 age,
\code{error}, the error of the C14 age and \code{depth}, the depth at which the observation was taken.

The model parameter can take one of the values bacon_simple, uniform_simple, sequential_simple, sequential_full, uniform_full.
}
\examples{
# run_model(sample_data, 'bacon_simple', modelled_points = 7, a_alpha = 140, b_alpha = 0.1, a_w = 5, b_w = 5, t_df = 3)
# run_model(MSB2K, 'bacon_simple', modelled_points = 21, a_alpha = 1.5, b_alpha = 0.075, a_2 = 5, b_w = 5, t_df = 3)

}
