% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{sample_data}
\alias{sample_data}
\title{Some sample age-depth data}
\usage{
sample_data()
}
\description{
\code{sample_data} returns a data frame of synthetic age-depth data which can be used as an input to the run_model function
}
