% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanagedepth-package.R
\docType{package}
\name{stanagedepth-package}
\alias{stanagedepth-package}
\alias{stanagedepth}
\title{The 'stanagedepth' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. https://mc-stan.org
}
